requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
certifi==2026.1.4
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
propcache==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.2.1
SQLAlchemy==2.0.46
typing_extensions==4.15.0
yarl==1.22.0
